Format all and fix some warnings and errors

+++ Prog 1/factorial.c
@@ -6,14 +6,13 @@ int main() {
 	printf("Enter a number: ");
 	scanf("%d", &num);
 
-	long int factorial = 1;
+	long long int factorial = 1;
 	int i = 1;
 
 	while (i <= num) {
 		factorial *= i;
 		i++;
 	}
-
 
 	printf("Factorial of %d is %lld", num, factorial);
